requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
matplotlib>=3.4.2
seaborn>=0.11.2
opencv-python>=4.5.3.56
Pillow>=8.2.0
tqdm>=4.61.1
albumentations>=1.0.0
pyyaml>=5.4.1
